moved utils to comman and +TODO
